feat:
Make test pass. Note that we found a defect in our test too.

It's valuable to set the expectation *before* you run the test.

I expected this test to pass and it failed. So that's a sign something's wrong.

And indeed - I saw @user instead of user in the test.

This is a smell but I'll leave it for now.

@practice:tdd @passing