test: increase coverage with end to end test
Turns out we'd made a mistake that wasn't caught by our existing unit test.

If a user wasn't activated, it still attempted to contact Castle.

So now we've changed our failing test to test for this behaviour to add coverage.

We can maybe convert this into a faster lower level test later.

@practice:tdd @passing @test_desiderata:inspires_confidence