wip: write high level test we want
We write the test we want.

What did we expect?

`#log_in_as` should have been unrecognised.

Instead, we get a strange error message:

```
Error:
LoginTest#test_attempting_login_shows_the_user's_welcome_page:
NameError: undefined local variable or method `session' for #<LoginTest:0x0000000105208a90>
    test/test_helper.rb:15:in `log_in_as'
    test/system/login_test.rb:6:in `block in <class:LoginTest>'
```

Why? Because we're coupled to the implementation provided by the superclass.

We can live with this for now by overriding this method with a private
method, but you can see the drawbacks of coupling.

It's not clear how to fix this quickly.

@antipattern:tight_coupling @design_metric:coupling
@design_practice:programming_by_intent