refactor: encapsulate what in helper methods
Got the test passing again.

We had to adjust our syntax so VCR could work.

But it's near the original intent.

We also added a trait to make tests more readable.

Note that once again, we're separating out the what from the how.

The test now doesn't know how the user should log in.

@design_practice:tdd__refactor @test_quality:readable
@design_principle:abstraction @design_principle:encapsulation