feat: notify Castle of failed login
We now have different VCR cassettes depending on if the users
password is valid or not.

This took a little adjusting to get right, but it shows the power
of encapsulation.

The production code changed, but none of the tests had to change.

The change was localised to just the login helper.

Note that we did tweak a fixture - this was a mistake we made earlier.

@design_metric:coupling @design_principle:abstraction
@design_heuristic:shoc__hides_internal_workings
@design_principle:encapsulation